Improve:Component,loadModelをまとめました。
